requirements.txt
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0

opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-requests==0.61b0
flask

